import error handling within adafruit_io.py
